requirements.txt
streamlit
pandas
psycopg2-binary
matplotlib
seaborn
numpy
google-generativeai
python-dotenv  